Complete replacing of Result with model object
Complete the replacing of `qiskit.Result` with the model-based class:
* move the remaining old methods of `_result` to the new class, with
  extra comments for deciding on the ones that make the public API.
* remove `qobj/_result.py` and `result/_result.py`.
* revise `result_from_old_style_dict`
