+++ qiskit/qobj/__init__.py
@@ -11,4 +11,3 @@ from ._qobj import (Qobj, QobjConfig, QobjExperiment, QobjInstruction,
                     QobjItem, QobjHeader, QobjExperimentHeader)
 from ._converter import qobj_to_dict
 from ._validation import validate_qobj_against_schema, QobjValidationError
-from ._result import Result, ExperimentResult
